move testimonial component out of vanilla folder
