+++ src/common/components/testimonial/testimonial.css
